Add optional threads argument
Passing a positive integer for the first argument now configures the
benchmark to run on multiple threads, if desired.

for example:

./avx_flops $(nproc)

No particular reason to add this, I was mostly just curious to see if
activating the AVX512 path caused higher power draw than just a simple
idle loop with the same thread count.

For reporting, the individual thread 'gflops' values are just summed
together, if running with more than one thread.